refactor: use `This` prefix for internal components
